Add multipage module
